feat: separate business logic from component
