+++ src/App.vue
@@ -1,10 +1,8 @@
 <template>
-  <EvidenceSearch :age="age"/>
+  <EvidenceSearch/>
 </template>
 
 <script setup lang="ts">
-import { ref } from 'vue';
-import EvidenceSearch from './components/EvidenceSearch.vue'
+import EvidenceSearch from './components/EvidenceSearch.vue';
 
-const age = ref(32);
 </script>
